fix(app): Clean Activities
Move all composable of Main and Login Activities to multi package for better code

issue #17
