+++ ui/src/main/java/com/despaircorp/ui/login/activity/LoginActivity.kt
@@ -46,6 +46,9 @@ import com.despaircorp.domain.real_estate_agent.model.RealEstateAgentEntity
 import com.despaircorp.shared.R
 import com.despaircorp.ui.login.LoginState
 import com.despaircorp.ui.login.LoginViewModel
+import com.despaircorp.ui.login.activity.composable.agent_drop_down.RealEstateAgentDropDown
+import com.despaircorp.ui.login.activity.composable.login.Login
+import com.despaircorp.ui.login.activity.composable.splash_screen.SplashScreen
 import com.despaircorp.ui.main.activity.MainActivity
 import com.despaircorp.ui.theme.RealEstateManagerKotlinTheme
 import com.despaircorp.ui.theme.merriweatherSans
@@ -117,190 +120,6 @@ fun LoginMain(modifier: Modifier = Modifier, viewModel: LoginViewModel) {
         }
     }
 }
-
-@Composable
-fun SplashScreen(modifier: Modifier = Modifier) {
-    Column(
-        modifier = modifier.fillMaxSize(),
-        horizontalAlignment = Alignment.CenterHorizontally,
-        verticalArrangement = Arrangement.Center,
-    ) {
-        Image(painter = painterResource(id = R.drawable.house), contentDescription = "")
-        
-    }
-    Column(
-        modifier = modifier.fillMaxSize(),
-        horizontalAlignment = Alignment.CenterHorizontally,
-        verticalArrangement = Arrangement.Bottom,
-    ) {
-        
-        LinearProgressIndicator(
-            modifier = Modifier
-                .padding(vertical = 20.dp),
-            color = MaterialTheme.colorScheme.surfaceVariant,
-            trackColor = MaterialTheme.colorScheme.outline,
-        )
-        
-        Text(
-            text = stringResource(id = R.string.app_name),
-            modifier = modifier.padding(vertical = 20.dp),
-            fontFamily = merriweatherSans,
-            fontWeight = FontWeight.Bold,
-            color = MaterialTheme.colorScheme.outline
-        )
-    }
-}
-
-@Composable
-fun Login(
-    modifier: Modifier = Modifier,
-    realEstateAgentEntities: List<RealEstateAgentEntity>,
-    onClick: (Int) -> Unit
-) {
-    
-    Column(
-        modifier = modifier.fillMaxSize(),
-        horizontalAlignment = Alignment.CenterHorizontally,
-        verticalArrangement = Arrangement.Center,
-    ) {
-        Image(painter = painterResource(id = R.drawable.house), contentDescription = "")
-        
-        Text(
-            text = stringResource(id = R.string.login_greetings),
-            modifier = modifier
-                .padding(vertical = 20.dp),
-            fontFamily = merriweatherSans,
-            fontWeight = FontWeight.Normal,
-            color = MaterialTheme.colorScheme.outline
-        )
-        
-        RealEstateAgentDropDown(realEstateAgentEntities, modifier, onClick)
-        
-    }
-}
-
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun RealEstateAgentDropDown(
-    realEstateAgentEntities: List<RealEstateAgentEntity>,
-    modifier: Modifier,
-    onClick: (Int) -> Unit
-) {
-    val text = stringResource(R.string.chose_someone)
-    var isExpanded by remember {
-        mutableStateOf(false)
-    }
-    
-    var agentChosen by remember {
-        mutableStateOf(text)
-    }
-    
-    var selectedIcon by remember {
-        mutableStateOf(R.drawable.gamer)
-    }
-    
-    var selectedIndex by remember { mutableStateOf(0) }
-    var selectedId by remember { mutableStateOf(0) }
-    
-    
-    Column(
-        modifier = modifier,
-        horizontalAlignment = Alignment.CenterHorizontally,
-        verticalArrangement = Arrangement.Center,
-    ) {
-        val activity = (LocalContext.current as? Activity)
-        
-        ExposedDropdownMenuBox(
-            expanded = isExpanded,
-            onExpandedChange = {
-                isExpanded = it
-            }
-        ) {
-            TextField(
-                value = agentChosen,
-                onValueChange = { },
-                readOnly = true,
-                leadingIcon = {
-                    Image(painter = painterResource(id = selectedIcon), contentDescription = "")
-                },
-                trailingIcon = {
-                    ExposedDropdownMenuDefaults.TrailingIcon(expanded = isExpanded)
-                },
-                colors = ExposedDropdownMenuDefaults.textFieldColors(),
-                modifier = modifier
-                    .menuAnchor()
-                    .focusProperties {
-                        canFocus = false
-                    },
-                
-                )
-            
-            ExposedDropdownMenu(
-                expanded = isExpanded,
-                onDismissRequest = {
-                    isExpanded = false
-                }
-            ) {
-                realEstateAgentEntities.forEachIndexed { index, item ->
-                    DropdownMenuItem(
-                        text = {
-                            Row(
-                                modifier = Modifier
-                                    .fillMaxSize(),
-                                verticalAlignment = Alignment.CenterVertically
-                            ) {
-                                Image(
-                                    painter = painterResource(id = item.imageResource),
-                                    contentDescription = "",
-                                    modifier = modifier.padding(8.dp)
-                                )
-                                Text(
-                                    text = item.name,
-                                    modifier = modifier.padding(8.dp),
-                                    fontFamily = merriweatherSans,
-                                    fontWeight = FontWeight.Normal,
-                                    color = MaterialTheme.colorScheme.outline
-                                )
-                            }
-                            
-                        },
-                        onClick = {
-                            selectedIndex = index
-                            isExpanded = false
-                            selectedIcon = item.imageResource
-                            agentChosen = item.name
-                            selectedId = item.id
-                        }
-                    )
-                }
-            }
-        }
-        Button(
-            modifier = Modifier
-                .padding(vertical = 24.dp),
-            onClick = {
-                if (selectedId == 0) {
-                    Toast.makeText(
-                        activity?.applicationContext,
-                        activity?.applicationContext?.getString(R.string.you_have_to_chose_a_profile_to_continue),
-                        Toast.LENGTH_SHORT
-                    ).show()
-                } else {
-                    onClick.invoke(selectedId)
-                }
-            }, colors = ButtonDefaults.buttonColors(
-                containerColor = MaterialTheme.colorScheme.secondary
-            )
-        ) {
-            Text(
-                stringResource(R.string.continue_button), fontWeight = FontWeight.Normal,
-                color = MaterialTheme.colorScheme.outline
-            )
-        }
-    }
-}
-
 
 @Preview(showBackground = true, uiMode = Configuration.UI_MODE_NIGHT_YES)
 @Composable
