+++ ui/src/main/java/com/despaircorp/ui/login/activity/composable/splash_screen/SplashScreenComposable.kt
@@ -1,0 +1,53 @@
+package com.despaircorp.ui.login.activity.composable.splash_screen
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.LinearProgressIndicator
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import com.despaircorp.shared.R
+import com.despaircorp.ui.theme.merriweatherSans
+
+
+@Composable
+fun SplashScreen(modifier: Modifier = Modifier) {
+    Column(
+        modifier = modifier.fillMaxSize(),
+        horizontalAlignment = Alignment.CenterHorizontally,
+        verticalArrangement = Arrangement.Center,
+    ) {
+        Image(painter = painterResource(id = R.drawable.house), contentDescription = "")
+        
+    }
+    Column(
+        modifier = modifier.fillMaxSize(),
+        horizontalAlignment = Alignment.CenterHorizontally,
+        verticalArrangement = Arrangement.Bottom,
+    ) {
+        
+        LinearProgressIndicator(
+            modifier = Modifier
+                .padding(vertical = 20.dp),
+            color = MaterialTheme.colorScheme.surfaceVariant,
+            trackColor = MaterialTheme.colorScheme.outline,
+        )
+        
+        Text(
+            text = stringResource(id = R.string.app_name),
+            modifier = modifier.padding(vertical = 20.dp),
+            fontFamily = merriweatherSans,
+            fontWeight = FontWeight.Bold,
+            color = MaterialTheme.colorScheme.outline
+        )
+    }
+}
